% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tema_personal_oscuro.R
\name{tema_personal_oscuro}
\alias{tema_personal_oscuro}
\title{Tema oscuro para graficos ggplot2, crea gráficos ideales para presentaciones}
\usage{
tema_personal_oscuro()
}
\description{
Tema oscuro para graficos ggplot2, crea gráficos ideales para presentaciones
}
\examples{
library(ggplot2,temaPersonal)
data <- data.frame(x = 1:10,y = c(2, 3, 5, 7, 11, 13, 17, 19, 23, 29))
ggplot(data, aes(x = x, y = y)) +
 geom_line(color = "yellow", linewidth = 1) +
 geom_point(size = 3,color="yellow") +
 labs(title = "Gráfico de Ejemplo",
      subtitle = "Uso del tema_personal en ggplot2",
      caption = "Datos de ejemplo") +
 tema_personal_oscuro()
}
