% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tema_personal_pastel.R
\name{tema_personal_pastel}
\alias{tema_personal_pastel}
\title{Tema para graficos  de pastel y donas
elimina los ejes de un grafico convencional}
\usage{
tema_personal_pastel()
}
\value{

}
\description{
Tema para graficos  de pastel y donas
elimina los ejes de un grafico convencional
}
\examples{
tema_personal_pastel(grafico_donus)
}
