% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpersonal_discreto.R
\name{ggpersonal_discreto}
\alias{ggpersonal_discreto}
\title{configurar graficos discretos,colores}
\usage{
ggpersonal_discreto(grafico)
}
\arguments{
\item{grafico}{object " Grafico a personalizar"}
}
\value{
object "Grafico configurado
}
\description{
configurar graficos discretos,colores
}
\examples{
library(ggplot2,temaPersonal)
theme_set(tema_personal())
data("mtcars")
grafico_barra = ggplot(mpg, aes(class)) + geom_bar(fill="tomato")
ggpersonal_discreto(grafico_barra)
}

