% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tema_personal_pastel.R
\name{tema_personal_pastel}
\alias{tema_personal_pastel}
\title{Tema para gráficos de tipo pastel y donas}
\usage{
tema_personal_pastel()
}
\description{
Función para aplicar tema a graficos de pastel, elimina los ejes de un gráfico convencional
}
\examples{
library(ggplot2,temaPersonal)
df = data.frame( categorias = c("cat1","cat2","cat3","cat4"), porcentaje = c(25,20,15,40))
ggplot(df,aes(x = "",y = porcentaje, fill = categorias))+
  geom_bar(stat = "identity",color="white")+
  coord_polar(theta="y") +
  tema_personal_pastel()
}
